refactor: Claudeレビュー指摘への根本的な対応完了
## 主な変更内容

### 1. parse_version関数の共通化 (推奨対応)
**問題**: 2つのファイルでparse_version関数が重複定義されていた
- scripts/migrate_metadata_v1_2_0.py:26-45
- scripts/migrate_metadata.py:211 (ローカル関数)

**解決策**: 共通ユーティリティモジュールに集約
- 新規作成: src/utils/version.py
  - parse_version()関数を定義
  - 詳細なdocstringと使用例を追加
- scripts/migrate_metadata_v1_2_0.py
  - 関数定義を削除してimportに変更
- scripts/migrate_metadata.py
  - ローカル関数を削除してimportに変更
  - エラーメッセージを明確化
- tests/test_migrate_metadata_v1_2_0.py
  - importを src.utils.version から取得するよう変更

**効果**:
- コードの重複を排除（DRY原則）
- 一箇所でのメンテナンスが可能
- 将来的な拡張が容易

### 2. マジックナンバーの定数化 (任意対応)
**問題**: GenderSumValidatorクラスでマジックナンバーが散在

**解決策**: クラス定数として定義
- _CONVERSION_TIMEOUT = 30  # subprocess timeout
- _TOLERANCE = 0.01  # 浮動小数点比較の許容誤差
- _MAX_ERROR_SAMPLES = 10  # エラーサンプルの最大数

**変更箇所**:
- src/validators/gender_sum_validator.py:23-29
  - 定数を定義
- 使用箇所を定数参照に変更:
  - Line 226: timeout=self._CONVERSION_TIMEOUT
  - Line 157: abs(expected - total_val) > self._TOLERANCE
  - Line 420: errors[:self._MAX_ERROR_SAMPLES]

**効果**:
- 設定値の変更が容易
- コードの可読性向上
- 定数の意味が明確

### 3. テストの修正
**問題**: parse_version関数の共通化によりエラーメッセージが変更された

**解決策**:
- tests/test_migrate_metadata.py のテストを修正
- 正規表現パターンを柔軟に (厳密なマッチから部分マッチへ)
  - r"Invalid version format: 'abc'"
  → r"Invalid version format.*'abc'"

### 4. コーディング規約準拠
- 全角括弧を半角括弧に修正 (RUF003)
  - scripts/migrate_metadata.py:211
  - 「（共通ユーティリティを使用）」→「(共通ユーティリティを使用)"

## テスト結果
- ✅ 全317テスト成功
- ✅ Ruff lint通過 (既存の許容される警告のみ)
- ✅ RUF002/RUF003チェック通過

## レビュー対応状況
✅ Claude推奨項目: parse_version関数の共通化
✅ Claude任意項目: マジックナンバーの定数化
✅ Claudeコーディング規約: 全角括弧の修正
