+++ tests/test_migrate_metadata.py
@@ -212,18 +212,18 @@ class TestMigrationRegistry:
     def test_compare_versions_invalid_format(self) -> None:
         """不正なバージョンフォーマットでValueErrorを発生させる."""
         # 数値以外を含むバージョン
-        with pytest.raises(ValueError, match=r"Invalid version format: '1\.a\.2'"):
+        with pytest.raises(ValueError, match=r"Invalid version format.*'1\.a\.2'"):
             MigrationRegistry.compare_versions("1.a.2", "1.0")
 
-        with pytest.raises(ValueError, match=r"Invalid version format: 'abc'"):
+        with pytest.raises(ValueError, match=r"Invalid version format.*'abc'"):
             MigrationRegistry.compare_versions("1.0", "abc")
 
         # 空文字列
-        with pytest.raises(ValueError, match=r"Invalid version format: ''"):
+        with pytest.raises(ValueError, match=r"Invalid version"):
             MigrationRegistry.compare_versions("", "1.0")
 
         # スペースを含む
-        with pytest.raises(ValueError, match=r"Invalid version format: '1 0'"):
+        with pytest.raises(ValueError, match=r"Invalid version"):
             MigrationRegistry.compare_versions("1 0", "1.0")
 
     def test_is_downgrade(self) -> None:
